Corrected instructions for Makefile flash: by removing TARGET
Previously, when a learner built the makefile by following the README, "make flash" generated an error. The issue was a reference to TARGET that didn't seem defined anywhere. (Maybe that was just me and this is invalid!) By making the change proposed, the makefile instructions align with the Makefile in Step-0-... and Step-1-... folders. Now "make flash" works for me with my main.c from the Minimal Firmware sections instructions and the Step-0-... main.c